references/backend/requirements.txt
fastapi==0.109.2
pydantic-settings==2.2.1
requests==2.31.0
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.3
celery[redis]==5.3.6
python-dotenv==1.0.1
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
email-validator==2.1.0.post1
pydantic[email]==2.6.1
pyotp==2.9.0
geoip2==4.8.0
user-agents==2.2.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0
sqlalchemy-utils==0.41.1
fastapi-users[sqlalchemy]==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
starlette==0.36.3
structlog==24.1.0
fastapi-limiter==0.1.6
fastapi-mail==1.4.1 